requirements.txt
pandas==1.4.3
netCDF4==1.5.8
Pillow==8.3.1
matplotlib==3.3.4
flask==2.0.2
scipy==1.6.1
s3fs==2023.1.0
numpy==1.0
